requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
langchain
python-dotenv
openai
wikipedia
langchain-community
fastapi-cors